change how infer-duration works
Specifically, I generalized Flags.can_cancel and Flags.cancel_next to work
on any concurrent note with the same instrument, and made
Flags.infer_duration processing independent of that.

This was all a big attempt to get kotekan calls with a final note to
cooperate with a normal call afterwards.  Ultimately it was a failure,
because it looks like it's a lot harder to get right than I thought.

But as a side effect, I made Derive.Call.Bali.Gangsa calls all consistently
omit the initial note and take an initial arg to add it back.  I also changed
cross-block cancellation to work by putting a Flags.cancel_next at the final
note, rather than a Flags.can_cancel at the first one.

I think this is generally better, but it's hard to be sure.