start to use Score.PControl for pitch controls, instead of Score.Control
I'm also using a null PControl for the default pitch, instead of Nothing.
This is less convenient in some ways, but overall I think it's cleaner.
It also makes it more natural to write calls that can take either a control
or a pitch control, differentiating with a leading '#'.

This is only a partial conversion since I'm still using Control in
PitchMap.