put Control and PControl into Typecheck
This means I can put them in the call signature and have the call mechanism
check and convert for me.

I'm not totally sure I like this, because it means that a pitch control
name needs a leading '#' even if it would otherwise be unambiguous, e.g.
(pitch-control '#name' (4c)).  But maybe the consistency is good?