change TimeVector.merge so it doesn't create duplicate samples
The previous implementation would do that if chunks went backwards in time,
which can happen if events are moved around, say by the 'd' call.

Unfortunately this way is less efficient because it reverses the list twice.
Hopefully it doesn't matter because the lists are usually very short.  The
new way also seems more regular in that later chunks always win, but it is
more complicated, so I'm not sure if it's really worth it just to get rid
of a warning that might not matter anyway.

I also added merge_left, even though I'm not using it, because it's simpler
and more efficient.  Unfortunately it doesn't seem to be as useful as
merge_right in practice.