implement sequence-rt
This turned out to be hard because I don't actually have the RealTime
duration of a block available anywhere, and in fact it's lost as soon as the
block is normalized to a duration of 1t.

This implementation records the RealTime end time of the call as it is
evaluated, which is too late to do any timing changes based on it, so it's
not as useful as the ScoreTime CallDuration.  I might have to change to a
similar implementation as CallDuration.