make nested tempo tracks all normalize to the end of the block
Previously they were just buggy and messed up the tempo.  I wound up with
a bit of refactoring and spending way too much time trying to understand all
the complicated stretching.  The dumb part is that I don't actually care about
this feature all that much, I just wanted to fix a bug.

However, nested tempos are still not normalized for the toplevel block and
should be, and are not implemented for hybrid or absolute tempo tracks.