add initial implementation of Derive.Call.India.Gamakam3
This is yet another attempt at writing down gamakam.  This approach is sort
of at a middle level between the low level Gamakam, which required multiple
transpose tracks, and the high level Gamakam3, which goes entirely on the note
track.  This one goes in the pitch track, but expects a parent pitch track
with the base swarams for each note.  Like Gamakam2, it implements its own
specialized vocabulary, but winds up recreating a simpler version of the Call
framework rather than reusing it.  I can't tell if that's a problem to fix, or
a feature where it can omit irrelevant details.

There are several unsolved problems.  One is that the parent pitch track has
to be repeated in the gamakam track for flat notes.  I should be able to solve
that with a interleave pitch ControlOp.  A more serious one is the usual
problem of getting the previous and next pitches.  This will likely require
some more extensive modifications and I'm still not sure if any of my plans
will work.  But if I get something reliable and general, I can get rid of the
two or three previous unsatisfactory attempts.