add ControlOp support to pitch tracks, and add interleave to both control and pitch
This was meant to solve a problem with the use of two pitch tracks in
Gamakam3, but as usual it just introduced another problem I didn't think
of.  That is that the signal winds up with extra samples at note end, which
presumably should have been trimmed after inversion.  I'm not sure exactly
how this happens but it makes me think interleave is maybe not a great
solution after all.