extend Gamakam3 calls to include dynamics
This involved redoing the parser with a real fancy parser and quite a bit of
refactoring.  'state_from_dyn' is not yet implemented.  I'm also not quite
sure this is enough, or is even a usable approach, but I think I'll probably
need something like it, so I can commit this for now while I experiment some
more.