add Args.lookup_next_logical_pitch and all the accompanying machinery
There winds up being quite a lot of complex support, including a new
derivation mode, special code in the default note deriver and Sub.inverting,
invoked via EvalTrack.derive_neighbor_pitches.

It seems like an awful lot of code scattered in a lot of places just so the
Gamakam3 call can get the next logical pitch.  So I consider it experimental
for now, and probably only worth it if the approach turns out to be useful in
general.  The other approach is to go postproc, which is what I've done for
other calls that need to know about the future.