replace UiMsg.UpdateZoom with UpdateTimeScroll
This should fix a bug where the zoom is unexpected reset after a zoom-to-fit.
I don't know exactly what the cause was, because it happens so rarely, but
I suspect that a race in Ui.fltk_event_loop causes the haskell State to get
an out-of-date Zoom when the UI is scrolled from the fltk side while zooming
from the haskell side.

I didn't fix the race, but since the UI actually never changes the zoom, I
can avoid it by not having the UI send the zoom factor at all, but only the
scroll value.  Ui.fltk_event_loop is now commented with a description of the
problem, so I could fix it if I see symptoms which I suspect are caused by it.