add Derive.Typecheck
This is the implementation for the new Typecheck class.  The difference is
that it can cause the type conversion to happen in Deriver, which means I
can convert scalar values to functions, and functions to scalar values.
Currently I do that conversion manually, which means I'm inconsistent about
it and it doesn't show up in the call doc.  With the new Typecheck I can
also coerce both ControlRef and ControlFunction to a single type, which was
the original motivation.

Since Typecheck no longer has a 1:1 relationship with Val constructors, I
also split 'to_val' off into its own class.