remove the global post-derive sort
This is basically a revert of 'sort events after derivation'.  The problem
is that I still need to keep events sorted because I want to efficiently
find neighbors, and in fact functions like Post.neighbors implicitly rely
on the list being ordered.

So I have to go back to keeping the stream in sorted order, which means to
be safe I have to sort it after every transformation.  I added the
Post.*asc variants back in so I can mark which postprocs expect to retain
the order, but I'll be introducing a safer and more efficient mechanism
soon.