change NoteDeriver return type from [LEvent.LEvent Score.Event] to a new Stream type
The original intent was to get some kind of handle on whether or not the
event list is sorted.  Stream is just a newtype, so it's not really an
improvement in that respect, but at least now that it has its own type I
can add features to it without having to change all those files again.
Also it makes it explicit where I'm trusting that the list is sorted
because of the Stream.from_sorted_list constructor.

One other benefit is that I can no longer combine streams with (++), but
I have to use (<>), which will do an ordered merge for Score.Events.