change how TrackWarps are collected, to be more efficient
Profiling showed that 'compare' from Derive.Stack was at the top, taking
6.7% time and it was all due to sorting the stacks before I collected
the TrackWarps in TrackWarp.collections.

Previously I would collect a TrackWarp for every track, but only have a
TrackId for tracks with the same warp as their parent.  But due to
inversion, every track winds up being every note, so there are tons of
redundant entries, each one with a stack.  Now, I only collect warps for
tempo tracks, and at the end I fill in the TrackIds of all the children.