add features to support read-only scores
The idea is to make it harder to accidentally change a score.  You can mark
a score read-only by removing the write permission.  The changes make the
saving functions properly handle a read-only file.

Also, now when there is no SaveFile you can't save at all, instead of it
inferring a name from the namespace and trying to overwrite that file.
If a git repo is readonly, it'll load but without a SaveFile, so it won't
try to write checkpoints to it.  A new 'x' indicator reminds when it's
not saving.