change selections to not clip when they go out of track bounds
I now have a special key to collapse the selection to a point, so I don't
need to mash it up because it went off the edge of the block.  This is also
consistent with timewise selection, which doesn't clip either.

As a side-effect, the point of a non-point selection is now the current
track and current pos, rather than min track and min pos, since otherwise
it could wind up on an invalid tracknum.  I think this is more consistent
anyway, and it should be visually apparent now that the cur end always has
an arrow.