remove dyn-atk
Unfortunately it never worked because it didn't merge with 'dyn' in the
right order.  This doesn't matter for commutative operations like 'mul',
but 'scale' is not commutative.

What's more, since it depends on the start times of notes, it can't
possibly be right since a signal track has no way to know about the note
tracks underneath it.

The general concept of a signal sampled only at note start still seems
useful, and is easy to do, but only if it doesn't need to be combined with
a continuous signal.