when creating a keymap binding, unshift all keys, not just letters
This means I can bind to '~' as a shorthand for shift + '`'.  I implemented
it because I wanted to bind some keys based on the symbol rather than
the physical location, but I wound up changing my mind about that because
they were awkward to type.  So I'm not actually using this yet, but maybe I
will someday.  Besides, the rules in Cmd.Keymap about what happens with
shift were ridiculously complicated and now they're simple, so I think this
is a good change just on that account.