reyong: fix a bug where aftertouch keyswitches would get lost
The problem was that I was using aftertouch for both damp level and
keyswitch, but only looking at the current state on a note on, so the one
before that was lost.

The solution was to entirely get rid of the on release damping and use
explicit mute notes.  This way I only use aftertouch for keyswitches, and
is a lot simpler besides.