shakefile: add functions to infer packages along with module imports
I thought I wound want this to integrate the Synth and give it separate
package dependencies, but it would have required me to also make all the
other things non-global, like test, profile, etc.  It seemed like a lot of
work for an unclear benefit.  Especially if I consider that part of the
reason to put it in the same repo is to get convenient global things like
tests.

So this just adds some unused functions and rearranges some other things.