Split Perform.Midi.Instrument.Instrument to Perform.Midi.Patch.Patch.
This makes Patch specifically for perform, and removes the awkward thing
where a Patch has an Instrument but some fields are needed only for perform
and have to be filled in by Convert, and the cost of a few duplicated
fields.  Also it reverses the names, so Patch is now the low level Perform
type, and Instrument is now the higher level one, though I haven't yet
renamed Instrument.Patch to Instrument.Instrument.

Also I moved some simple types from Instrument.Inst to Instrumnet.InstTypes
to reduce dependencies.