rename Perform.Midi.Patch to Perform.Midi.Types, move Perform.Event into it
Mostly this is so I cane rename Perform.Midi.Instrument to Patch, but also
means modules that just want to deal with MIDI perform events don't have to
import Perform.

I agonized about the naming quite a bit, and I don't love Types very much,
but there's a lot of precedent for this name for a module of low level type
definitions.