add support for randomization within the notes generated by a single call
Since the random seed is based on the stack, notes generated by the same
call would all have the same stack, and hence they would all be
"randomized" in exactly the same way.

This adds a serial number to notes which should be incremented for each
Score.Event produced by the same UI event.  The serial number is then
added to the seed, so every note should be randomized uniquely but still
deterministically.