Cmd.MidiThru: rely on the environ and controls from the Performance's Dynamic
Previously, the thru function would look up the instrument and apply its
environ and controls itself.  I'm not sure why I added this, because it
seems to be redundant, given that they should already be in the Dynamic,
and a more accurate representation of the particular values at that point
in the score.  Specifically, this also takes into account any transforms
at the global, block, or track level.

As before, though, I still zero out any transposers other than octave,
since from experience it's very confusing when one pitch is entered into
the score but another one is played via MIDI thru.