remove Cmd.Cmd.Cmd type synonym
It was non-polymorphic which meant some cmds couldn't be called from tests,
e.g. via CmdTest.  I could have added the monad parameter, but just deleted
it instead.  It wasn't used that often and while it did abstract the
signature of a generic cmd, it didn't seem to be so valuable as to be worth
the extra level of indirection.