add Derive.Call.StaticMacro
This is a statically-bound variant of Derive.Call.Macro.

I don't know if I really need all this, since maybe a plain dynamic macro
is enough, but it seems nice to get the arguments exported to the macro's
documentation.  And this is kind of an automatic version of what I've
written a number of times by hand so... useful I guess?