fix a bug where the deriver doesn't revert the state when there is an exception
When Derive.Deriver.Internal.local changes the state, it doesn't reset it
on an exception, because it expects that whoever catches the exception
will do that.  That turns out to be EvalTrack.run_derive... who in fact
didn't revert to the old state.  Now it does.

This is a giant bug in a really fundamental part of the deriver and I'm
amazed it's taken this long to notice.  This meant that all manner of old
state would stick around if an event threw an exception.  I guess it's
because I don't write tests with both failing and succeeding calls, and
when I'm playing a score I generally stop if there is an error.  I only
noticed because it messed up the stack on subsequent errors.