gamakam: simplify down arguments work for dyn calls
Now all dyn calls will look for a single digit argument.  This means I can
no longer just type a digit to move to that dyn, so now it's an argument
for a '.' call.  I think this is ok because I don't really do a lot of move
to dyn like I do move to pitch.