add support to retune instruments with PitchedKeymaps via MIDI or ksp
For some reason this wound up being a huge hassle when it really should
have been straightforward.

I initially put the pitch ranges into the Patch.Scale itself and had a
function to warp the Scale into [(Key, NoteNumber)] but with the Keys
matching the pitch ranges, but it was always clunky and awkward.
Eventually I reverted the whole approach in favor of the Scale always
just being the instrument-independent scale, and combining that with an
AttributeMap to create either absolute or relative tuning information.
The problem is that for whatever reason trying to keep straight which
values represented which kinds of pitches.  For example, I originally had
PitchedKeymap with low and high Keys, but then a NoteNumber for the base
pitch.  This is wrong because the pitch range starts with with the address
of a pitch (i.e. a Key) while the actual pitch depends on the tuning.

Maybe it's legitimately complicated, or maybe my brain is just too small.