solkattu: remove the karvai concept
It greatly complicated matters when I wanted to add speed and nadai
changes.  But on patterns that use it it's a lot easier to write than
doing the duration math, so I might want to bring it back again, but
hopefully in a way that doesn't complicate everything so much.