add 'nruk' call
This was surprisingly involved.  First I wanted to be able to have the
start and end speeds have different time types, not for any really concrete
reason, but because it seemed like the sort of thing that should be easy to
do.  Turns out it wasn't and I had to change Speed to use durations instead
of speeds.  That in turn led to a separate Speed type to differentiate from
Duration.

Then I discovered that despite having both Call.Macro and Call.StaticMacro,
I still have no way to make an alias like `nruk | o`, where the arguments
are dispatched to a statically bound `nruk`, and `o` is dynamically bound.
I worked around it but maybe I should extend StaticMacro to be able to do
that.