create a new socket on app startup, and delete on quit
This is instead of unconditionally deleting any socket and creating a new
one on startup.  The previous behaviour lead to an unreachable `seq` when
I accidentally started a new instance and it deleted the old instance's
port.

As a bonus, I can now communicate with multiple `seq` instances though I
doubt it's a very useful feature.