add yet another mechanism for getting next pitch, Args.lookup_next_pitch
This one is implemented by having each note track evaluate a pitch track
underneath it and stashing the results in Dynamic.  The previous
NotePitchQuery technique of doing the evaluation on demand, and doing it
for entire calls rather than just pitch calls, made it really confusing
once it got into recursive calls.

If this works out, I can get rid of NotePitchQuery and replace it with this
much simpler version.