kotekan: add norot2, which uses the new Args.lookup_next_pitch
The old norot emits its preparation before the event start, canceling out
any notes already there.  This one emits the preparation for the next note.

I think both ways need manual tweaking.  The old way frequently needs a
flag to turn of the prepare when the norot is meant to start on the beat,
and the new way needs to turn it off when the next note isn't a norot.
I think the new way is a bit less confusing though, because its notes are
all within its extent, and it doesn't need a complicated mechanism to
suppress other notes.