simplify how delayed args work
They used to have an attr to indicate that an arg was present, and then
the arg itself.  Since only the attr was removed by the realize call, it
events confusingly looked as if they had not yet been processed.  Now there
is only an arg entry in Score.state_delayed_args, which is deleted when
it's processed.