change negative event implementation to use a flag instead of negative duration
This means the event start time, and hence the EventMap index, is at the
min edge of the event instead of the end for negative events.  This in turn
means I can have a negative event that ends where a positive one starts,
which is something that I would like to be able to express.

Negative events still use negative durations in the fltk code, and for tests,
simply because it's still a convenient syntax.  However it's definitely hard
to read negative events like (10, -8, "abc") because it changes the start
time, so I might change that while there still aren't too many tests with
them.

I was hoping to get rid of all of the special case code in Ui.Events and
Cmd.Edit for negative events, and I still could if I want commands like
set_duration to work physically (set the bottom of the event) rather than
logically.  But working logically still seems to make sense and I already had
all that code so I figured I might as well keep it.  It turned out I had to
actually rewrite lots of stuff and I'm not so sure it was worth it it, but it
works now so let's hope it was.

Also I modified Derive.Control.trim_signal to allow a final control signal on
slices for a negative note event.  This meant I had to put the event
orientation into TrackTree.Track, and due to the extra option, I merged
track_sliced and track_inverted into a Sliced type.