try to fix Edit.toggle_zero_timestep for negative events, but fail
It turns out Events.insert and Events.remove are wrong for negative events.
Or at least wrong from a certain point of view.  The problem is that
Ui.Events wants to treat negative events as end-weighted with respect to
times and ranges, but can't do it entirely because the whole point of
the Orientation change was that events are now always indexed by
Event.start.  So I now think this inconsistency is too hard and I should
just treat all events the same.