take all special case Event.Negative handling out of Ui.Events
It used to try to treat them as if the start is at the end, where in fact it
used to be, and logically reverse all of the range oriented functions.
Unfortunately, because now events are always indexed at Event.start it was
never going to be really consistent, and was really complicated besides.

So I'm moving to just treating events as they are, with the start at the
start, even if negative ones draw with a trigger line and text at the
bottom.  As an experiment, Edit.cmd_set_duration is still logical, but the
rest of the commands will become physical.

I also replaced Edit.cmd_toggle_zero_duration with cmd_toggle_zero_timestep
and simplified it to use ModifyEvents.selection_expanded.