add types for CallName, ArgName, and Doc
I wanted to differentiate the Text args that call constructors tend to take.
However, I'm not sure if this is really necessary, since it's not like I had
a big problem mixing those up.  However, though the change was a bit of a
pain, once it's already done it doesn't seem like much work to satisfy,
so maybe it's generally good.

Since CallNames and ArgNames are used for defaulting, it's good to notice
when I'm constructing them, and hence could have illegal characters in them.