fltk: add fltk/FloatingInput.cc and use it for the event floating input
This is a separate borderless window instead of being an overlapping
widget.  The main reason is that now it can be wider than the track, even
if that would mean going off the edge of the window, but as a nice
side-effect it means I can get rid of various hacks to support an
overlapping widget that changes size.