make various MIDI instrument settings configurable per-score
Initially I just wanted to be able to override flags for instrument
allocations, but in the process this turned into a huge endeavour.

The essential change is that now per-score settings are in a Settings
record, which the Patch has in patch_defaults and the Patch.Config has
in config_settings.  When an allocation is created, patch_defaults is
copied to config_settings and can subsequently be modified.

This is more expressive and more regular than the previous way, where
fields like Scale were duplicated in Patch and Config, and
Cmd.lookup_instrument would merge them, giving preference to Config, and
put the results back into Patch.  It seemed confusing that the Patch was
modified by the lookup function, so it's nice that that's gone now.

The main source of complication is that since Config now has an absolute
config rather than a relative "merge me with Patch" config, I now need
the Patch to create a Config.  This is a pain for tests.  Most of them
don't modify Settings though, so I added DeriveTest.SimpleAllocations
and various functions which look up the Settings from the instrument db.
Also, it turns out the various allocation templates in Local.Instrument.*
still really want to be relative changes, so those now emit an "incomplete"
Config, where incompleteness is flagged with a bogus pitch bend range, and
must be merged with 'patch_defaults'.  Obviously that's pretty gross and
I don't like it but it seemed like a lot less hassle than a duplicate
"relative" Settings.

I'm a bit unhappy about the increased complexity, and quite unhappy about
the size of the change for a seemingly simple feature.  Several times I
considered reverting the whole thing, but in the end it seemed like maybe
some of the complexity was due to happen anyway as I added more locally
configurable fields.