repl: implement a way for the repl to edit a string in vi
There is a new Edit Result which has the string to edit, and a function
to call with the edited result.  The idea is that it should be a get/set
pair, and now I can edit internal strings with vi.  That in turn means I
could store a local ky file which is saved along with the score, and maybe
use it to replace the global transform.