add Cmd.Instrument.Bali.wrap, which wraps pitches to fit in, say, calung or jegogan range
Initially there was a bug where transposition was applied twice, which
revealed that I don't track transposition status for Pitch.Pitch or
Pitch.Note, so it's ambiguous whether or not I should clear out
transposers.  I opted to assume that Pitch and Note are always
post-transpose, but it's another awkward thing about the transposition
system.