change '=' syntax to allow an optional merge operator
Now I can write `x = 5 sub` as `x=-5`.  The idea is that this will
encourage me to merge properly, e.g. it's easy to have a bug like
`%t-dia = -1` which will not cooperate with an existing transposition.

I don't like so much that the order is backwards from usual, e.g. `=*`
instead of `*=`, but necessary since I allow any character but space or `=`
in the first word.  The other thing is that `x=-2` means `x =- 2`, so if
you want to set to -2, you have to put a space in.  This latter one I think
is a bigger problem, but the only solutions I could think of, like using
`=~` for subtraction, seem to be even more error-prone.