fltk: move damage range tracking from OverlayRuler to Track
This reduces some duplication because now range damage and clipping is all
done in the track draw() methods, but adds some duplication because now
RulerTrack and EventTrack have to handle selection and selection damage
instead of dispatching to OverlayRuler.

Ultimately I'm doing this so I can split off the selection handling to a
OverlaySelection, so I can then draw the selection on top of the event
triggers, which are on top of the ruler overlay.  It seemed convoluted to
try to do that when OverlaySelection was simultaneously handling ruler,
selection, and damage range.