fltk: fix a bug where event tracks didn't get redrawn properly
Actually it turns out that damage area for events was always broken, but
it was hidden by the fact that as a child widget, the RulerOverlay was
getting FL_DAMAGE_ALL and redrawing everything anyway.  Not totally sure
why though.

In any case, I no longer try to optimize event updates at all anymore, and
instead just redraw everything regardless of the damaged range.  In
practice this is what I was always doing anyway.