automatically add Stack.Serial from Derive.state_event_serial
Since the Stack is now unique, I can replace the separate serial in
CacheKey and separately adding it to the random seed.  I'm not sure why
I didn't do this from the beginning.

Of course it changes the randomization yet again.  I should come up with
some way to stop doing that.