rework how negative events are handled
This completely reworks negative events, yet again.  There are two somewhat
orthogonal changes.

The first is that negative events are back to using negative duration
instead of an explicit orientation field.  This means their start time is
now at the end of the event, and the end time can be less than the start
time.  This is how things used to be, but I switched to a separate
orientation field and always positive durations under the impression that
it would #1, make the awkward negative event special casing disappear, and
the time in practice.  However, #1 was basically a failure because it turns
out that regardless of how the events are encoded, it's logically the most
intuitive for editing cmds to act as if the duration is negative, and it
turns out the explicit orientation scheme sometimes makes them more
convenient, but also sometimes less convenient, for no real gain in
simplicity.  In fact it led to inconsistent behaviour where the start of a
negative event is logically at its bottom, but certain editing cmds edit
from the top since that's where the start really is, and must be, to
support #2.  #2 did work, but it turns out it made it impossible to express
a zero-dur positive event against a negative event like |<--|, which does
in fact happen.

So, I want back to negative durations, and addressed #2 by keying the
events by (start, orientation) instead of just start.  This also has the
bonus effect of allowing back-to-back zero duration events, for which I can
already think of some uses.  It also will hopefully resolve the issue where
inverting notes moves zero-dur controls together, which causes one to be
destroyed.

That leaves the problem of how to address edits to those coincedent events,
which was the second change, which was to add an orientation to the
selection.  At the fltk level, this gives explicit control over the
direction of the arrows on the selection, analogous to the arrows on
events.  At the haskell level, the selection orientation disambiguates
which event you mean to edit.

I'm not sure if this will turn out to be awkward or confusing in practice.
It does resolve some other ambiguities, such as which event duration to
extend when there are two options, as in |--> <--|, and maybe other uses
will turn up.

Going back to negative durations meant that yet again I wound up rewriting
all the annoying finicky cmds in Cmd.Edit.  However, it led to a certain
amount of refactoring and hopefully they're slightly better structured than
before.  Default selection behaviour is now more complicated since it
attempts to reconcile positive and negative orientations, but hopefully
will be more consistent and intuitive in the end.  Also such a major change
led to numerous small refactorings in many places.