solkattu: in MridangamDsl, use NoSollu with an explicit stroke instead of a StrokeMap
This way the Sollkattu don't have confusing dummy sollus.  I also put a
hack in Pretty Solkattu to omit NoSollu when there is an explicit stroke.

Originally I tried to generalize Solkattu to replace Sollu with a Stroke
directly, but then I can't do a cancel_karvai.  Then I tried to generalize
the Sollu part, but then I'd have to rewrite the functions in Korvai, and
the types are already kind of too complicated.  This way is simpler and
almost as good.