im: change the signal type to Util.TimeVector, change Time to RealTime
This adds more sharing of types between the sequencer and im systems.
Originally I thought to make im completely independent, but since putting
them in the same source repo I'm thinking there's nothing wrong with
sharing some basic types.

This gets rid of the expensive conversion step.  I still have to serialize
to disk, so I could have merged it into that, but I'd need to make sure the
on-disk representation is identical, and the safest way to do that is to
share the types.