solkattu: add support for light and heavy emphasis
I wrapped the stroke type in a Realize.Stroke, which can hold additional
attributes.  It turned out to be surprisingly complicated.  Also, it turns
out ToCall is not good enough because it's only a CallId and emphasis might
require transforms and hence a whole expression.  For the moment I force
it into CallId which will work for LSol but no one else.