implement support for explicit signal discontinuities
This means that signals are assumed to have interpolation between each
sample, so a discontinuity must be represented explicitly as two coincident
samples.

This breaks signal interleave, which gamakam4 relied on, so it's disabled
while I figure out what to do about that.

Ultimately it's because the im backend does sample-level interpolation,
while for MIDI I just rely on whatever the MIDI synthesizer happens to
hardcode.