change ParseTitle to use attoparsec, and add support for track call annotations
Instead of pretending like the first section of a track title is a list of
vals, I just treat it as its own language with its own parser.  Track call
!-syntax added another special case.  The advantage is that I can more
easily add special syntax now, but the error messages get worse.

A side-effect is that I had to reverse control merge syntax from
'merge control' to 'control merge'.  The postfix notation is more
consistent so I don't mind this so much.