TrackCall: make a pitch track call be exclusive with the name
The idea is that the track call sets the control name itself by returning
one.

I'm not sure I like this because it winds up being the wrong type for
pitch tracks, and irrelevant for note tracks, and seems orthogonal to
track calls anyway, but I'll have to think about it more, and meanwhile
this is consistent with control track syntax.