fix an assertion failure when opening a floating input on the last track
It was caused by the 0 <= tracknum < tracks fix, since
TrackTile::floating_open didn't get tracknum-1 as usual.  It turns out that
was because I used to send the tracknum in the UiMsg.  I don't do that
anymore, so I could get rid of the awkward exception where floating_open's
tracknum is at +1.