solkattu: get rid of the explicit stroke parameter in Solkattu.NoteT
The idea was that I hardly ever use it, and tags can do the same kind of
thing, but in a more instrument-independent way.

This took way longer than it should have, and the fly in the ointment was
that MridangamScore used explicit strokes.  So I generalized NoteT on sollu
instead, which is a nicer solution to NoSollu + explicit strokes everywhere.
I had tried this before, but failed due to some awkward details like
cancel_karvai, and having to put two type parameters on everything.  At some
point in the past I fixed the cancel_karvai problem by adding Solkattu.NoteT,
and getting rid of 'stroke' made me more willing to add a 'sollu'.

Unfortunately the integration with MridangamScore is still not good, since
I need to pass 'mridangam_stroke' instead of 'mridangam' handling both stroke
and sollu scores, but I'll try to figure that out later.