lilypond: fix buggy combination of attach and emit calls
The problem was that attach would try to put code on non-note code events,
and that emit code calls seemed like they worked as notes, but actually
didn't since they emit 0 duration code events.  The solution was to enforce
that emit calls are either transforms or 0 duration, and have no sub notes.