write a MonadError instance for DeriverM
I meant it so I could use tryRight etc. instead of Derive.require_right,
but it's not actually useful for that because of course the error type
isn't Text.  But I'll leave the instance there since it doesn't hurt, and
will save me some time the next time I want to make that mistake.