lilypond: refactor tuplets so that the lilypond backend handles them
Previously, the tracklang level tuplet call would run a mini-lilypond
conversion and embed the results in a \tuplet { ... } code block.
Unfortunately, this disrupted the usual linear evaluation sequence, which
made modal articulations like pizz/arco not work, and slurs not get attached.

I changed it so that the tracklang level tuplet emits the unstretched notes
along with a special event that tells the lilypond level to gather them into
a tuplet.

Unfortunately it got really complicated and confusing along the way, since
I have to bypass the usual meter rules (the old way got that on accident,
since its mini-conversion would sort of ignore the meter), and all of the
timing is just really hard to think about.  I also spent a significant amount
of time and several dead-ends trying to get nested tuplets to work... and
they *still* don't work, but I'm out of energy for now.

The main goal was to not lose articulation signs in tuplets, and that seems
to be accomplished, so good enough for now.