solkattu: add the initial group implementation
The idea is that when I drop sollus from a sequence, it remembers what
they were, so realize can use this context to figure out the strokes.
Otherwise for the common case of a reducing pattern, I have to put all the
suffix or prefix fragments in the stroke map.

Also, I hope to extend this to encompass patterns, so that they will behave
as normal sollus.

The Notation.drop functions are broken with regard to nested groups, and
this breaks a bunch of tests, but the patch is already huge as it is, so
I'll check it in and fix the problems later.