lilypond: make wide trills emit tremolo notation
I also integrated attribute trill into the standard trill, so it can fall
back on emitting the notes itself for a trill which is wider than a whole
step.

Also, now that the 'tr' call works properly for lilypond, I removed +trill
from Process.simple_articulations, so +trill is used solely as a
performance indication and dosen't wind up in notation, while 'tr' is for
both.