lilypond: sort the pitches in a chord
glissandoMap refers to chord notes by index, which means I need them to be
in a predictable order.

This change was way more annoying than it should have been.  I changed
Types.event_pitch from Text to a new Pitch type and all the tests had to
change.  I could have left pitch as Text, but it felt dumb to re-parse the
pitch just for sorting.