lilypond: implementa a new tie-breaking heuristic
This one is a bit more lenient for non-binary meters like 3/4.  The key
difference is that exact matches are preferred over ties, along with
a special hack so non-duple meters break at rank-1 except exact matches
that put higher rank in the middle.

It's actually not that complicated but it took a long time to get there.
And still it's not really right, but it may be time to start adding special
cases instead of further tweaking the algorithm.  The current algorithm
feels increasingly ad-hoc and hacky too.