BlockResize: add a feature to splice the ruler into the top block
Figuring out how to do this is mostly what caused this whole feature to be
so annoying in the first place.

I actually wound up adding two versions, the original ruler splicing one
in LBlock.{add,remove}_time_ruler and a simpler one in
LBlock.{add,remove}_time which assume the top ruler is already right and
push it down with no splicing.  The latter is simpler and expresses music
moving without the meter changing, while the former expresses that added
time also moves the meter beyond it.