lilypond: refactor conversion to use explicit rests, and generalize code positions
This started as a refactor to explicitly represent rests as Events, instead
of noticing gaps between Events and filling them in with lilypond rests.
The idea was that the complicated code to generate rests and then mix code
events into it could be replaced with something to globally mix code into
all events, with no special cases for rests.  All of the special casing for
rests and code mixing was getting complicated as special tremolo and tuplet
handling were added.  Also, there were various bugs and inconsistencies in
0 duration code-only events vs. attaching code directly to notes, namely
that the 0 duration events could only support prepend and append.

Along the way, I changed the code position to support all the logical
permutations of [note, chord] [prepend, append] [first, last, all], instead
of a bunch of ad-hoc instances.

Mixing in code events globally led to the realization that free-standing
code can only support chord-prepend-first and chord-append-first by its
very nature, since it's not attached to a note, so can't know which note's
ties to distribute over.  So free-standing code events still only support
prepend and append, but at least it's represented statically as a separate
type now.

This was supposed to be a simplification, but overall
Perform.Lilypond.Process increased by 100 lines.  Because of course it did.
At least 0 dur code events should now attach to rests at the correct time.